move linechart to custom common
