+++ custom/ewsghana/reports/__init__.py
@@ -11,11 +11,12 @@ from corehq.apps.products.models import SQLProduct
 from corehq.apps.programs.models import Program
 from corehq.apps.reports.datatables import DataTablesHeader
 from corehq.apps.reports.generic import GenericTabularReport
-from corehq.apps.reports.graph_models import LineChart, MultiBarChart, PieChart
+from corehq.apps.reports.graph_models import MultiBarChart, PieChart
 from corehq.apps.hqwebapp.decorators import use_nvd3
 from custom.ewsghana.filters import EWSRestrictionLocationFilter
 from corehq.apps.reports.standard import CustomProjectReport, ProjectReportParametersMixin, DatespanMixin
 from custom.common import ALL_OPTION
+from custom.common.models import LineChart
 from custom.ewsghana.filters import ProductByProgramFilter, EWSDateFilter
 from custom.ewsghana.models import EWSExtension
 from memoized import memoized
